Add branch_protection to repos serviced by Zuul
In order for post jobs (e.g. for deployments) to work properly and to
avoid them running on development branches, Zuul needs to be configured
to exclude unprotected branches. So we need to configure branch
protection for the branches that Zuul actually should act upon. The
current tooling allows to do this only for the default_branch of a repo,
which is good enough for now.

This patch adds a branch protection template that only sets a minimal
set of restrictions, but that is good enough for Zuul to recognize the
main branch as protected. This template is applied to all repos that are
currently known to Zuul.

Update the gitcontrol submodule with the fixes needed to handle this data.

Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>